Generacion Qr en reservas activas

+++ lib/pages/client/reservas_realizadas/reservation_ended.dart
@@ -64,7 +64,7 @@ class ReservasFinalizadasCliente extends StatelessWidget {
               model: data['vehiculo']['marcaVehiculo'],
               plate: data['vehiculo']['placaVehiculo'],
               status: data['estado'],
-              total: data['total'],
+              total: data['total'].toDouble(),
               typeVehicle: data['vehiculo']['tipo'],
               id: document.id,
               idDuenio: data['parqueo']['idDuenio'],
